Don't follow Bootstrap in zeroing `border-radius` for `.input-group .select2-container--bootstrap:not(:first-child):not(:last-child) .select2-selection`. Reverts the CSS changes from 26c88a12acdfa22f3c762f423fffb5a1c36494cf
When Select2 is used in context of a Bootstrap input group and only an appended button is present, the original, hidden `<select>` element will be causing .select2-container--bootstrap to not be `:first-child`, causing Select2's border-radius to be removed (which is not what we want). The existing `.select2-bootstrap-prepend` and `.select2-bootstrap-append` classes provide an easy fix for the "appended and prepended button"-case where border-radius should indeed be nulled.